+++ src/main/java/kusitms/backend/chatbot/application/service/ChatbotApplicationService.java
@@ -1,15 +1,24 @@
-package kusitms.backend.chatbot.application;
+package kusitms.backend.chatbot.application.service;
 
+import kusitms.backend.chatbot.application.factory.ClovaRequestFactory;
+import kusitms.backend.chatbot.application.factory.MessageFactory;
+import kusitms.backend.chatbot.application.dto.request.ChatbotRequestDto;
+import kusitms.backend.chatbot.application.dto.response.GetClovaChatbotAnswerResponseDto;
 import kusitms.backend.chatbot.domain.enums.*;
-import kusitms.backend.chatbot.dto.response.GetGuideChatbotAnswerResponseDto;
+import kusitms.backend.chatbot.application.dto.response.GetGuideChatbotAnswerResponseDto;
+import kusitms.backend.chatbot.domain.service.ChatbotApiClient;
 import kusitms.backend.chatbot.status.ChatbotErrorStatus;
 import kusitms.backend.global.exception.CustomException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import reactor.core.publisher.Mono;
 
 @Service
 @RequiredArgsConstructor
-public class ChatbotService {
+public class ChatbotApplicationService {
+    private final ChatbotApiClient chatbotApiClient;
+    private final ClovaRequestFactory clovaRequestFactory;
+    private final MessageFactory messageFactory;
 
     /**
      * 주어진 카테고리 이름, 경기장 이름, 질문 번호에 따라 답변을 조회합니다.
@@ -70,5 +79,20 @@ public class ChatbotService {
                 )
                 .map(answer -> GetGuideChatbotAnswerResponseDto.of(answer.getAnswers(), answer.getImgUrl()))
                 .orElseThrow(() -> new CustomException(ChatbotErrorStatus._NOT_FOUND_GUIDE_CHATBOT_ANSWER)); // 답변 없을 때 예외
+    }
+
+    /**
+     * Clova 챗봇 답변을 가져옵니다.
+     *
+     * @param message 사용자가 보낸 메시지
+     * @return Clova 챗봇 답변을 포함하는 Mono 객체
+     * @throws IllegalArgumentException 메시지가 null 또는 빈 문자열인 경우
+     */
+    public Mono<GetClovaChatbotAnswerResponseDto> getClovaChatbotAnswer(String message) {
+        ChatbotRequestDto request = clovaRequestFactory.createClovaRequest();
+        request.getMessages().add(messageFactory.createUserMessage(message));
+
+        return chatbotApiClient.requestChatbot(request)
+                .map(GetClovaChatbotAnswerResponseDto::of);
     }
 }
